requirements.txt
streamlit
opencv-python>=4.5.5.64
numpy>=1.14.0
requests>=2.27.1
Pillow>=5.2.0
tempfile
ios
time
sqlalchemy
fastapi
pydantic
uvicorn
facenet-pytorch
torch
torchvision
logging
python-dotenv
jose[bcrypt]
bcrypt
json
pandas>=0.23.4
gdown>=3.10.1
tqdm>=4.30.0
tensorflow>=1.9.0
keras>=2.2.0
Flask>=1.1.2
flask_cors>=4.0.1
mtcnn>=0.1.0
retina-face>=0.0.1
fire>=0.4.0
gunicorn>=20.1.0
